messages/requirements.txt
SQLAlchemy==1.3.22
Flask==1.1.2
waitress==2.0.0
python-dotenv==0.17.0
Flask-SQLAlchemy == 2.4.4
PyJWT==2.1.0
click==7.1.2